Finalise merge of `origin/dev1.6` into `data_loading`
