+++ hoki/tests/test_load.py
@@ -137,23 +137,23 @@ class TestLoadAllRates(object):
     @patch("hoki.load.model_output")
     def test_load_rates(self, mock_model_output):
         mock_model_output.return_value = self.data
-        x = load.all_rates(f"{data_path}", "imf135_300"),\
+        x = load.rates_all_z(f"{data_path}", "imf135_300"),\
             "The rates cannot be initialised."
 
     # Load rates
     with patch("hoki.load.model_output") as mock_model_output:
         mock_model_output.return_value = data
-        x = load.all_rates(f"{data_path}", "imf135_300")
+        x = load.rates_all_z(f"{data_path}", "imf135_300")
 
     # Test wrong inputs
     def test_file_not_present(self):
         with pytest.raises(AssertionError):
-            _ = load.all_rates(f"{data_path}", "imf135_300"),\
+            _ = load.rates_all_z(f"{data_path}", "imf135_300"),\
                 "The file is not present, but the load function runs."
 
     def test_wrong_imf(self):
         with pytest.raises(HokiKeyError):
-            _ = load.all_rates(f"{data_path}", "i"),\
+            _ = load.rates_all_z(f"{data_path}", "i"),\
                 "An unsupported IMF is taken as an input."
 
     # Test output
@@ -187,7 +187,7 @@ class TestLoadAllSpectra(object):
         # Set the model_output to the DataFrame
         mock_model_output.return_value = self.data.to_numpy()
         mock_isfile.return_value = True
-        spec = load.all_spectra(f"{data_path}", "imf135_300")
+        spec = load.spectra_all_z(f"{data_path}", "imf135_300")
 
         # Check if compiled file is created
         assert os.path.isfile(f"{data_path}/all_spectra-bin-imf135_300.npy"),\
@@ -202,7 +202,7 @@ class TestLoadAllSpectra(object):
 
     def test_load_pickled_file(self):
 
-        spec = load.all_spectra(f"{data_path}", "imf135_300")
+        spec = load.spectra_all_z(f"{data_path}", "imf135_300")
 
         # Check output numpy array
         npt.assert_allclose(
@@ -229,7 +229,7 @@ class TestLoadAllEmissivities(object):
         # Set the model_output to the DataFrame
         mock_model_output.return_value = self.data.to_numpy()
         mock_isfile.return_value = True
-        res = load.all_emissivities(f"{data_path}", "imf135_300")
+        res = load.emissivities_all_z(f"{data_path}", "imf135_300")
 
         # Check if compiled file is created
         assert os.path.isfile(f"{data_path}/all_ionizing-bin-imf135_300.npy"),\
@@ -244,7 +244,7 @@ class TestLoadAllEmissivities(object):
 
     def test_load_pickled_file(self):
 
-        res = load.all_emissivities(f"{data_path}", "imf135_300")
+        res = load.emissivities_all_z(f"{data_path}", "imf135_300")
 
         # Check output numpy array
         npt.assert_allclose(
